third_party/blink/web_tests/external/wpt/css/css-position/sticky/position-sticky-left-and-right-overconstrained-expected.txt

This is a testharness.js-based test.
[FAIL] start of scroll container
  assert_equals: expected 20 but got 200
[FAIL] right before the in-flow position of the sticky box
  assert_equals: expected 180 but got 200
Harness: the test ran to completion.
